# Arm64 CRC32 parallel computation optimization for RocksDB