@siying 2019-07-15
We don't have bandwidth in reviewing the actual algorithm, but since crc32c_test passes, I assume the algorithm at least generates correct CRC values. I will wait for some people to chime in from portability aspects. If nobody objects, I can merge.

CC @siddontang @yiwu-arbug in case you are interested.