@yiwu-arbug 2019-07-16
@siying the PR will help with our testing with ARM, but unfortunately we are not familiar with CRC algorithm either. thanks for the heads up!