@siddontang 2019-07-17
Thanks @siying 

Unfortunately, as @yiwu-arbug said before, we don't write code on ARM, but we can help you do some benchmarks in TiKV after this PR merged.